requirements.txt
numpy==1.24.3
Pillow==11.1.0
streamlit==1.40.1
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb

